functions/extract/requirements.txt
google-api-python-client
functions-framework
google-cloud-secret-manager
google-cloud-storage
requests
duckdb
feedparser
pandas
beautifulsoup4